Fix iPhoneX layout, + small changes
